pyproject.toml
[project]
name = "edge-c2-simulator"
version = "0.1.0"
description = "Multi-domain C2 simulation engine for Edge C2 demonstration"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "aiohttp>=3.9",
    "websockets>=12.0",
    "pyais>=2.6",
    "shapely>=2.0",
    "geopy>=2.4",
    "click>=8.1",
    "uvicorn>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "aioresponses>=0.7",
]

[project.scripts]
edge-c2-sim = "scripts.run_simulator:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["simulator*", "scripts*"]
